pyproject.toml
[tool.poetry]
name = "tcp-server"
version = "0.1.0"
description = "Servidor TCP com suporte a múltiplas conexões"
authors = ["Seu Nome <seu.email@exemplo.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^1.0.0"
taskipy = "^1.14.0"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
start = "python -m tcp_server.server"
test = "pytest tests/ -v"
lint = "black . && isort ."
coverage = "pytest --cov=tcp_server tests/"